Added support for named fields that contain the '.' characters
Its part of the format() syntax to allow the getattr syntax, which make names like 'hello.world' possible. Previously it would try to use the named field as a regex group name, which did not work as anything but a-zA-Z0-9_ seem to be forbidden characters.

The issue is solved by temporarily replacing the forbidden characters with a character sequence that is not, and converting the group name back once the result is prepared.

A test was added to verify the feature