Fix test_too_many_fields with Python 3.5
Python versions before 3.5 had a limit of 100 groups in regular
expressions. This limit was removed during 3.5 development:

http://bugs.python.org/issue22437
https://hg.python.org/cpython/rev/0b85ea4bd1af

The test_too_many_fields test asserts that the limit exists by
attempting to parse a string with 15 fields, which triggers the 100
named groups limit.

Adjust the test so that if first checks to see whether the limit of 100
named groups exists, and only assert that parsing 15 fields fails if
that is the case.