Support 'evaluate_result' flag on parse(), search() and findall() methods
If the ``evaluate_result`` flag is set to ``False`` a ``Match`` instance
is returned instead of a ``Result`` instance.

A ``Match`` instance has one method ``evaluate_result()`` which returns
a ``Result`` instance for this very match.

This feature is especially useful if you want to lazy evaluate the
custom type conversion function but check if it matches some string in
advanced.

Signed-off-by: Timo Furrer <timo.furrer@roche.com>