Clarify error message on invalid format spec
The error message I got this morning was a little bit cryptic:

>>> ValueError: type ' 5>d' not recognised

I suggest changing it to be a little clearer. "type not recognized" seems to be saying that "d" isn't a valid type. However the problem in my case was just had inadvertently swapped two characters (the correct format spec is " >5d").