Don't use `import *` to introduce the module
Awesome package! Using wildcard imports (`import *`) obfuscates which parts of the example code come from the module, and it violates the suggestions in [PEP 8](https://www.python.org/dev/peps/pep-0008/). The fact that your implementation of `__all__` is motivated by name conflicts, and not API design, is exactly why wildcard imports are usually discouraged. I imported the module as `pe`, partially because `p` was already taken in your examples, and partially because it looks like `re` ;-p.